# openapi.yaml
openapi: 3.1.0
info:
  title: Official Document Delivery Framework API
  version: 0.1.0
  summary: API for email-linked event handling, instruction management, chat-driven document drafting, and document export.
  description: |
    This API supports:
    - ingesting forwarded emails into email events
    - managing document-writing instructions
    - chatting within an email event context
    - exporting generated official documents

servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://staging-api.example.com/v1
    description: Staging

tags:
  - name: Inbound Email
  - name: Email Events
  - name: Document Instructions
  - name: Chat
  - name: Documents

security:
  - bearerAuth: []

paths:
  /inbound/emails:
    post:
      tags: [Inbound Email]
      summary: Ingest a forwarded email
      description: |
        Internal/system-facing endpoint used when a user forwards an email to the platform mailbox.
        The backend creates or updates an email event, optionally using title as an initial correlation hint.
      operationId: ingestForwardedEmail
      security:
        - ingestApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InboundEmailCreateRequest'
      responses:
        '201':
          description: Email event created or matched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailEvent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /email-events:
    get:
      tags: [Email Events]
      summary: List email events
      operationId: listEmailEvents
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: title
          in: query
          description: Filter by event title
          schema:
            type: string
        - name: status
          in: query
          description: Filter by event status
          schema:
            $ref: '#/components/schemas/EmailEventStatus'
      responses:
        '200':
          description: Email events returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailEventListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /email-events/{emailEventId}:
    get:
      tags: [Email Events]
      summary: Get an email event
      operationId: getEmailEvent
      parameters:
        - $ref: '#/components/parameters/EmailEventId'
      responses:
        '200':
          description: Email event returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailEvent'
        '404':
          $ref: '#/components/responses/NotFound'

  /email-events/{emailEventId}/chat-history:
    get:
      tags: [Chat]
      summary: Get chat history for an email event
      operationId: getChatHistory
      parameters:
        - $ref: '#/components/parameters/EmailEventId'
      responses:
        '200':
          description: Chat history returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatHistory'
        '404':
          $ref: '#/components/responses/NotFound'

  /email-events/{emailEventId}/messages:
    post:
      tags: [Chat]
      summary: Send a chat message within an email event
      description: |
        Mail is treated as chat in this workflow. A user message can ask the system to:
        - update the generated official document
        - draft an email requesting more information
      operationId: createChatMessage
      parameters:
        - $ref: '#/components/parameters/EmailEventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageCreateRequest'
      responses:
        '200':
          description: Assistant reply generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatReply'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /email-events/{emailEventId}/export:
    post:
      tags: [Documents]
      summary: Export a document for an email event
      description: Generate and store a document derived from the email event and current chat context.
      operationId: exportDocument
      parameters:
        - $ref: '#/components/parameters/EmailEventId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportDocumentRequest'
      responses:
        '201':
          description: Document exported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentExportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /document-instructions:
    get:
      tags: [Document Instructions]
      summary: List document-writing instructions
      operationId: listDocumentInstructions
      responses:
        '200':
          description: Instructions returned successfully
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DocumentInstruction'
    post:
      tags: [Document Instructions]
      summary: Create a document-writing instruction
      operationId: createDocumentInstruction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentInstructionCreateRequest'
      responses:
        '201':
          description: Instruction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentInstruction'
        '400':
          $ref: '#/components/responses/BadRequest'

  /document-instructions/{instructionId}:
    get:
      tags: [Document Instructions]
      summary: Get a document-writing instruction
      operationId: getDocumentInstruction
      parameters:
        - $ref: '#/components/parameters/InstructionId'
      responses:
        '200':
          description: Instruction returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentInstruction'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Document Instructions]
      summary: Update a document-writing instruction
      operationId: updateDocumentInstruction
      parameters:
        - $ref: '#/components/parameters/InstructionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentInstructionUpdateRequest'
      responses:
        '200':
          description: Instruction updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentInstruction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Document Instructions]
      summary: Delete a document-writing instruction
      operationId: deleteDocumentInstruction
      parameters:
        - $ref: '#/components/parameters/InstructionId'
      responses:
        '204':
          description: Instruction deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'

  /documents/{documentId}:
    get:
      tags: [Documents]
      summary: Get document metadata
      operationId: getDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Document metadata returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          $ref: '#/components/responses/NotFound'

  /documents/{documentId}/download:
    get:
      tags: [Documents]
      summary: Download an exported document
      operationId: downloadDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Document binary content
          content:
            application/pdf: {}
            application/vnd.openxmlformats-officedocument.wordprocessingml.document: {}
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ingestApiKey:
      type: apiKey
      in: header
      name: X-Ingest-Key

  parameters:
    EmailEventId:
      name: emailEventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    InstructionId:
      name: instructionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  schemas:
    EmailEventStatus:
      type: string
      enum:
        - received
        - in_progress
        - waiting_for_information
        - ready_for_export
        - exported
        - closed

    DocumentFormat:
      type: string
      enum:
        - pdf
        - docx
        - txt

    Email:
      type: object
      required:
        - id
        - sender
        - title
        - content
        - date
      properties:
        id:
          type: string
          format: uuid
        sender:
          type: string
          format: email
        title:
          type: string
        content:
          type: string
        date:
          type: string
          format: date-time

    InboundEmailCreateRequest:
      type: object
      required:
        - sender
        - title
        - content
        - date
      properties:
        sender:
          type: string
          format: email
        title:
          type: string
        content:
          type: string
        date:
          type: string
          format: date-time
        externalMessageId:
          type: string
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'

    Attachment:
      type: object
      required:
        - fileName
        - mediaType
        - url
      properties:
        fileName:
          type: string
        mediaType:
          type: string
        url:
          type: string
          format: uri

    EmailEventSummary:
      type: object
      required:
        - id
        - title
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          $ref: '#/components/schemas/EmailEventStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        latestDocumentId:
          type: string
          format: uuid

    EmailEvent:
      allOf:
        - $ref: '#/components/schemas/EmailEventSummary'
        - type: object
          required:
            - sourceEmail
          properties:
            sourceEmail:
              $ref: '#/components/schemas/Email'
            correlationKey:
              type: string
              description: Optional title-based correlation key used during ingestion.

    EmailEventListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmailEventSummary'
        nextCursor:
          type: string

    DocumentInstruction:
      type: object
      required:
        - id
        - title
        - content
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    DocumentInstructionCreateRequest:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
        content:
          type: string

    DocumentInstructionUpdateRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
      additionalProperties: false

    ChatRole:
      type: string
      enum:
        - user
        - assistant
        - system

    ChatChannel:
      type: string
      enum:
        - web
        - email

    ChatMessage:
      type: object
      required:
        - id
        - emailEventId
        - role
        - channel
        - content
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        emailEventId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/ChatRole'
        channel:
          $ref: '#/components/schemas/ChatChannel'
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    ChatMessageCreateRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        intent:
          type: string
          enum:
            - update_document
            - draft_information_request_email
            - general
        relatedDocumentId:
          type: string
          format: uuid

    ChatHistory:
      type: object
      required:
        - emailEventId
        - messages
        - historicDocuments
      properties:
        emailEventId:
          type: string
          format: uuid
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
        historicDocuments:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSummary'

    SuggestedAction:
      type: object
      required:
        - type
        - label
      properties:
        type:
          type: string
          enum:
            - export_document
            - draft_email
            - request_more_information
        label:
          type: string

    ChatReply:
      type: object
      required:
        - replyMessage
      properties:
        replyMessage:
          $ref: '#/components/schemas/ChatMessage'
        suggestedActions:
          type: array
          items:
            $ref: '#/components/schemas/SuggestedAction'

    DocumentSummary:
      type: object
      required:
        - id
        - title
        - format
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        format:
          $ref: '#/components/schemas/DocumentFormat'
        status:
          type: string
          enum:
            - generating
            - ready
            - failed
        createdAt:
          type: string
          format: date-time

    Document:
      allOf:
        - $ref: '#/components/schemas/DocumentSummary'
        - type: object
          required:
            - emailEventId
            - updatedAt
          properties:
            emailEventId:
              type: string
              format: uuid
            updatedAt:
              type: string
              format: date-time
            version:
              type: integer
              minimum: 1
            downloadUrl:
              type: string
              format: uri

    ExportDocumentRequest:
      type: object
      properties:
        format:
          $ref: '#/components/schemas/DocumentFormat'
        title:
          type: string
        includeChatHistory:
          type: boolean
          default: true

    DocumentExportResponse:
      type: object
      required:
        - document
      properties:
        document:
          $ref: '#/components/schemas/Document'

    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true