# src/main/java/com/officialpapers/api/dto/DocumentExportResponse.java
package com.officialpapers.api.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class DocumentExportResponse {

    private DocumentDto document;
}