# src/main/java/com/officialpapers/api/dto/InboundEmailCreateRequest.java
package com.officialpapers.api.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.List;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class InboundEmailCreateRequest {

    @NotBlank
    @Email
    private String sender;

    @NotBlank
    private String title;

    @NotBlank
    private String content;

    @NotNull
    private LocalDateTime date;

    private String externalMessageId;

    private List<AttachmentDto> attachments;
}